+++ terraform/variables.tf
@@ -12,60 +12,85 @@ terraform {
 }
 
 variable "hetzner_token" {
-  default = "YOUR_HETZNER_CLOUD_TOKEN"
+  description = "Hetzner Cloud Token"
+  type        = string
 }
 
 variable "hetznerdns_token" {
-  default = "YOUR_HETZNER_DNS_TOKEN"
+  description = "Hetzner DNS Token"
+  type        = string
 }
 
 variable "location" {
-  # Hetzner Cloud data center, available are: FI Helsinki (hel1), DE Falkenstein (fsn1), DE Nuremberg (nbg1), US Ashburn (ash) and US Hillsboro (hil)
-  default = "hel1"
+  description = "Hetzner Cloud data center, available are: FI Helsinki (hel1), DE Falkenstein (fsn1), DE Nuremberg (nbg1), US Ashburn (ash) and US Hillsboro (hil)"
+  type        = string
+  default     = "hel1"
 }
 
 variable "os_type" {
-  default = "ubuntu-22.04"
+  description = "OS image type for created VMs"
+  type        = string
+  default     = "ubuntu-22.04"
 }
 
-variable "ssh_key_file" {
-  default = "~/.ssh/id_rsa.pub"
+variable "ssh_private_key_file" {
+  description = "Location of SSH Private Key on local notebook"
+  type        = string
+  default     = "~/.ssh/id_rsa"
+}
+
+variable "ssh_public_key_file" {
+  description = "Location of SSH Public Key on local notebook"
+  type        = string
+  default     = "~/.ssh/id_rsa.pub"
 }
 
 variable "user" {
-  # executing user on notebook
-  default = "georg"
+  description = "Executing user on notebook"
+  type        = string
+  default     = "georg"
 }
 
 variable "domain" {
-  default = "example.com"
+  description = "Used domain for IDE deployment"
+  type        = string
+  default     = "example.com"
 }
 
 variable "ide_ha_setup" {
-  # Variable to decide for IDE Non-HA or IDE HA setup, set to true for IDE HA setup
-  default = false
+  description = "Variable to decide for IDE Non-HA or IDE HA setup, set to true for IDE HA setup"
+  type        = bool
+  default     = false
 }
 
 # To scale IDE change below values
 
 variable "server_type_security" {
-  default = "cpx11"
+  description = "Server type of security server"
+  type        = string
+  default     = "cpx11"
 }
 
 variable "server_type_hub" {
-  default = "cpx31"
+  description = "Server type of hub server"
+  type        = string
+  default     = "cpx31"
 }
 
 variable "server_type_master" {
-  default = "cpx41"
+  description = "Server type of master server"
+  type        = string
+  default     = "cpx41"
 }
 
 variable "server_type_worker" {
-  default = "cpx51"
+  description = "Server type of worker server"
+  type        = string
+  default     = "cpx51"
 }
 
-# TTL for created A-records
-
 variable "a_record_ttl" {
-  default = 600
+  description = "TTL for created A-records in Hetzner Cloud"
+  type        = number
+  default     = 600
 }